pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "livecell-distillation"
version = "0.1.0"
description = "Cross-architecture knowledge distillation for single-cell classification on LIVECell."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Philip Graemer", email = "philip.graemer@strath.ac.uk"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
    "h5py>=3.0",
    "numpy>=1.21",
    "Pillow>=9.0",
    "tqdm>=4.60",
    "scikit-learn>=1.0",
    "matplotlib>=3.5",
    "seaborn>=0.12",
]

[project.optional-dependencies]
flops = ["ptflops>=0.7"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
include = ["livecell_distillation*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py39"
line-length = 100
